Reorganize docs.  Add glimmer docs.
