+++ app/controllers/ember.js
@@ -1,0 +1,6 @@
+import Controller from '@ember/controller';
+import Yaml from 'emblem-site/yaml/ember';
+
+export default class EmberController extends Controller {
+  yaml = Yaml
+}
