+++ app/controllers/glimmer.js
@@ -1,0 +1,6 @@
+import Controller from '@ember/controller';
+import Yaml from 'emblem-site/yaml/glimmer';
+
+export default class GlimmerController extends Controller {
+  yaml = Yaml
+}
