--- app/controllers/idioms.js
@@ -1,6 +1,0 @@
-import Controller from '@ember/controller';
-import Idioms from 'emblem-site/yaml/idioms';
-
-export default class IdiomsController extends Controller {
-  idiomsYaml = Idioms
-}
