+++ app/controllers/mustache.js
@@ -1,0 +1,6 @@
+import Controller from '@ember/controller';
+import Yaml from 'emblem-site/yaml/mustache';
+
+export default class MustacheController extends Controller {
+  yaml = Yaml
+}
