+++ app/controllers/syntax.js
@@ -1,6 +1,6 @@
 import Controller from '@ember/controller';
-import Syntax from 'emblem-site/yaml/syntax';
+import Yaml from 'emblem-site/yaml/syntax';
 
 export default class SyntaxController extends Controller {
-  syntaxYaml = Syntax
+  yaml = Yaml
 }
