+++ app/templates/components/render-yaml-docs.embl
@@ -5,13 +5,13 @@
 
   .grid-x
     .small-12.medium-4.large-3.cell.left-side
-      h3 = doc.title
+      h5 = doc.title
 
       = if doc.ember
         h5 Ember!
 
       = each (split-into-paragraphs doc.docs) as |paragraph|
-        p = paragraph
+        p == parse-markdown paragraph
 
     .small-12.medium-8.large-9.cell.right-side
       = if doc.emblem
