--- app/templates/idioms.embl
@@ -1,1 +1,0 @@
-= render-yaml-docs yaml=idiomsYaml
