+++ package.json
@@ -29,6 +29,7 @@
     "ember-truth-helpers": "2.0.0",
     "emblem": "^0.10.0",
     "js-yaml": "^3.12.0",
+    "markdown": "^0.5.0",
     "prettier": "^1.15.3"
   },
   "devDependencies": {
